+++ application/index/controller/Comment.php
@@ -11,29 +11,37 @@ class Comment extends Base
     // 发表评论
     public function addComment(Request $request)
     {
-        $status = 0;
-        $message = '发表失败';
-        $data = $request->post();
-        $data['username'] = $this->username;
-        $result = CommentModel::insertGetId($data);
-        if ($result != 0) {
-            $status = 1;
-            $message = '发表成功';
+        if ($request->isAjax()) {
+            $status = 0;
+            $message = '发表失败';
+            $data = $request->post();
+            $data['username'] = $this->username;
+            $result = CommentModel::insertGetId($data);
+            if ($result != 0) {
+                $status = 1;
+                $message = '发表成功';
+            }
+            return ['status' => $status, 'message' => $message, 'id' => $result];
+        } else {
+            return $this->error('非法访问');
         }
-        return ['status' => $status, 'message' => $message,'id'=>$result];
     }
 
     // 删除评论
     public function delComment(Request $request)
     {
-        $status = 0;
-        $message = '删除失败';
-        $data = $request->post();
-        $result = CommentModel::destroy($data);
-        if ($result == true) {
-            $status = 1;
-            $message = '删除成功';
+        if ($request->isAjax()) {
+            $status = 0;
+            $message = '删除失败';
+            $data = $request->post();
+            $result = CommentModel::destroy($data);
+            if ($result == true) {
+                $status = 1;
+                $message = '删除成功';
+            }
+            return ['status' => $status, 'message' => $message];
+        } else {
+            return $this->error('非法访问');
         }
-        return ['status' => $status, 'message' => $message];
     }
 }
